+++ test/Deferred.mjs
@@ -2,20 +2,19 @@
 
 /**
  * Deferred promise that can be externally resolved or rejected.
- * @template T
+ * @template [Resolves=void] What the promise resolves.
  */
 export default class Deferred {
   constructor() {
-    /**
-     * The promise.
-     * @type {Promise<T>}
-     */
-    this.promise = new Promise((resolve, reject) => {
-      /** Resolves the promise. */
-      this.resolve = resolve;
+    /** The promise. */
+    this.promise = /** @type {Promise<Resolves>} */ (
+      new Promise((resolve, reject) => {
+        /** Resolves the promise. */
+        this.resolve = resolve;
 
-      /** Rejects the promise. */
-      this.reject = reject;
-    });
+        /** Rejects the promise. */
+        this.reject = reject;
+      })
+    );
   }
 }