+++ changelog.md
@@ -16,6 +16,7 @@
 - Simplified dev dependencies and config for ESLint.
 - Fixed issues with GraphQL result related types from `types.mjs`.
 - Improved various JSDoc descriptions.
+- Improved the types relating to the `Deferred` class used in tests.
 
 ## 17.0.0